requirements.txt
csv

python-dotenv
scrapy